requirements.txt
boto3==1.28.57
Pillow>=10.1.0
requests
numpy
torch
json